requirements.txt
python-dotenv==0.15.0
tweepy==3.10.0
git+https://github.com/Binance-docs/Binance_Futures_python.git
